+++ rentACar-main/src/main/java/com/etiya/rentACar/business/concretes/ColorManager.java
@@ -3,6 +3,11 @@ package com.etiya.rentACar.business.concretes;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.etiya.rentACar.business.constants.messages.BrandMessages;
+import com.etiya.rentACar.business.constants.messages.ColorMessages;
+import com.etiya.rentACar.core.utilities.business.BusinessRules;
+import com.etiya.rentACar.core.utilities.results.ErrorResult;
+import com.etiya.rentACar.entities.Brand;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -44,24 +49,54 @@ public class ColorManager implements ColorService{
 
 	@Override
 	public Result save(CreateColorRequest createColorRequest) {
+		Result result = BusinessRules.run(checkExistingColor(createColorRequest.getColorName()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(createColorRequest, Color.class);
 		this.colorDao.save(color);
-		return new SuccessResult("Color added.");
+		return new SuccessResult(ColorMessages.add);
 	}
 
 
 	@Override
 	public Result delete(DeleteColorRequest deleteColorRequest) {
+		Result result = BusinessRules.run(checkExistingColorId(deleteColorRequest.getColorId()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(deleteColorRequest, Color.class);
 		this.colorDao.delete(color);
-		return new SuccessResult("Color added.");
+		return new SuccessResult(ColorMessages.delete);
 	}
 
 
 	@Override
 	public Result update(UpdateColorRequest updateColorRequest) {
+		Result result = BusinessRules.run(checkExistingColor(updateColorRequest.getColorName()), checkExistingColorId(updateColorRequest.getColorId()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(updateColorRequest, Color.class);
 		this.colorDao.save(color);
-		return new SuccessResult("Color added.");
+		return new SuccessResult(ColorMessages.update);
+	}
+
+	private Result checkExistingColor(String colorName1) {
+		String lowerCaseColorName = colorName1.toLowerCase();
+		for (Color color : colorDao.findAll()) {
+			String lowerCaseExistingColorName = color.getColorName().toLowerCase();
+			if(lowerCaseColorName.equals(lowerCaseExistingColorName)) {
+				return new ErrorResult(ColorMessages.duplicationError);
+			}
+		}
+		return new SuccessResult();
+	}
+
+	private Result checkExistingColorId(int colorId){
+		if (colorDao.existsById(colorId)){
+			return new SuccessResult();
+		}
+		return new ErrorResult(ColorMessages.colorIdNotFound);
 	}
 }
