+++ rentACar-main/src/main/java/com/etiya/rentACar/business/concretes/CorporateCustomerManager.java
@@ -3,6 +3,15 @@ package com.etiya.rentACar.business.concretes;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.etiya.rentACar.business.abstracts.UserService;
+import com.etiya.rentACar.business.constants.messages.CorporateCustomerMessages;
+import com.etiya.rentACar.business.constants.messages.CreditCardMessages;
+import com.etiya.rentACar.business.constants.messages.UserMessages;
+import com.etiya.rentACar.core.utilities.business.BusinessRules;
+import com.etiya.rentACar.core.utilities.results.*;
+import com.etiya.rentACar.entities.IndividualCustomer;
+import org.apache.commons.lang3.StringUtils;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,10 +21,6 @@ import com.etiya.rentACar.business.request.corporateCustomerRequests.CreateCorpo
 import com.etiya.rentACar.business.request.corporateCustomerRequests.DeleteCorporateCustomerRequest;
 import com.etiya.rentACar.business.request.corporateCustomerRequests.UpdateCorporateCustomerRequest;
 import com.etiya.rentACar.core.utilities.mapping.ModelMapperService;
-import com.etiya.rentACar.core.utilities.results.DataResult;
-import com.etiya.rentACar.core.utilities.results.Result;
-import com.etiya.rentACar.core.utilities.results.SuccessDataResult;
-import com.etiya.rentACar.core.utilities.results.SuccessResult;
 import com.etiya.rentACar.dataAccess.abstracts.CorporateCustomerDao;
 import com.etiya.rentACar.entities.CorporateCustomer;
 @Service
@@ -23,12 +28,14 @@ public class CorporateCustomerManager implements CorporateCustomerService{
 
 	private ModelMapperService modelMapperService;
 	private CorporateCustomerDao corporateCustomerDao;
+	private UserService userService;
 	
 	@Autowired
-	public CorporateCustomerManager(ModelMapperService modelMapperService, CorporateCustomerDao corporateCustomerDao) {
+	public CorporateCustomerManager(ModelMapperService modelMapperService, CorporateCustomerDao corporateCustomerDao, UserService userService) {
 		super();
 		this.modelMapperService = modelMapperService;
 		this.corporateCustomerDao = corporateCustomerDao;
+		this.userService = userService;
 	}
 
 	@Override
@@ -41,29 +48,60 @@ public class CorporateCustomerManager implements CorporateCustomerService{
 
 	@Override
 	public Result save(CreateCorporateCustomerRequest createCorporateCustomerRequest) {
+		Result result = BusinessRules.run(userService.existsByEmail(createCorporateCustomerRequest.getEmail()),
+				checkIfTaxNumberIsNumeric(createCorporateCustomerRequest.getTaxNumber()));
+		if (result != null){
+			return result;
+		}
 		CorporateCustomer corporateCustomer = modelMapperService.forRequest().map(createCorporateCustomerRequest, CorporateCustomer.class);
 		this.corporateCustomerDao.save(corporateCustomer);
-		return new SuccessResult("Corporate customer added.");
+		return new SuccessResult(CorporateCustomerMessages.add);
 	}
 
 	@Override
 	public Result delete(DeleteCorporateCustomerRequest deleteCorporateCustomerRequest) {
+		Result result = BusinessRules.run(checkIfUserIdExists(deleteCorporateCustomerRequest.getUserId()));
+		if (result != null){
+			return result;
+		}
 		CorporateCustomer corporateCustomer = modelMapperService.forRequest().map(deleteCorporateCustomerRequest, CorporateCustomer.class);
 		this.corporateCustomerDao.delete(corporateCustomer);
-		return new SuccessResult("Corporate customer deleted.");
+		return new SuccessResult(CorporateCustomerMessages.delete);
 	}
 
 	@Override
 	public Result update(UpdateCorporateCustomerRequest updateCorporateCustomerRequest) {
+		Result result = BusinessRules.run(checkIfUserIdExists(updateCorporateCustomerRequest.getUserId()),
+				checkIfTaxNumberIsNumeric(updateCorporateCustomerRequest.getTaxNumber()));
+		if (result != null){
+			return result;
+		}
 		CorporateCustomer corporateCustomer = modelMapperService.forRequest().map(updateCorporateCustomerRequest, CorporateCustomer.class);
 		this.corporateCustomerDao.save(corporateCustomer);
-		return new SuccessResult("Corporate customer updated.");
+		return new SuccessResult(CorporateCustomerMessages.update);
 	}
 
 	@Override
 	public CorporateCustomer getCustomerByCustomerId(int customerId) {
 		CorporateCustomer corporateCustomer = corporateCustomerDao.getById(customerId);
 		return corporateCustomer;
+	}
+
+	private Result checkIfUserIdExists(int userId){
+		CorporateCustomer corporateCustomer = corporateCustomerDao.getByUserId(userId);
+		if (corporateCustomer == null){
+			return new ErrorResult(UserMessages.userDoesNotExist);
+		}
+		return new SuccessResult();
+	}
+
+	public Result checkIfTaxNumberIsNumeric(String taxNumber){
+		if(StringUtils.isNumeric(taxNumber)){
+			return new SuccessResult();
+		}
+		else {
+			return new ErrorResult(CorporateCustomerMessages.invalidTaxNumberFormat);
+		}
 	}
 
 }
