+++ rentACar-main/src/main/java/com/etiya/rentACar/business/concretes/IndividualCustomerManager.java
@@ -3,6 +3,11 @@ package com.etiya.rentACar.business.concretes;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.etiya.rentACar.business.abstracts.UserService;
+import com.etiya.rentACar.business.constants.messages.IndividualCustomerMessages;
+import com.etiya.rentACar.business.constants.messages.UserMessages;
+import com.etiya.rentACar.core.utilities.business.BusinessRules;
+import com.etiya.rentACar.core.utilities.results.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,10 +17,6 @@ import com.etiya.rentACar.business.request.individualCustomerRequests.CreateIndi
 import com.etiya.rentACar.business.request.individualCustomerRequests.DeleteIndividualCustomerRequest;
 import com.etiya.rentACar.business.request.individualCustomerRequests.UpdateIndividualCustomerRequest;
 import com.etiya.rentACar.core.utilities.mapping.ModelMapperService;
-import com.etiya.rentACar.core.utilities.results.DataResult;
-import com.etiya.rentACar.core.utilities.results.Result;
-import com.etiya.rentACar.core.utilities.results.SuccessDataResult;
-import com.etiya.rentACar.core.utilities.results.SuccessResult;
 import com.etiya.rentACar.dataAccess.abstracts.IndividualCustomerDao;
 import com.etiya.rentACar.entities.IndividualCustomer;
 
@@ -24,11 +25,13 @@ public class IndividualCustomerManager implements IndividualCustomerService{
 	
 	private IndividualCustomerDao individualCustomerDao;
 	private ModelMapperService modelMapperService;
+	private UserService userService;
 	@Autowired
-	public IndividualCustomerManager(IndividualCustomerDao individualCustomerDao, ModelMapperService modelMapperService) {
+	public IndividualCustomerManager(IndividualCustomerDao individualCustomerDao, ModelMapperService modelMapperService, UserService userService) {
 		super();
 		this.individualCustomerDao = individualCustomerDao;
 		this.modelMapperService = modelMapperService;
+		this.userService = userService;
 	}
 
 	@Override
@@ -41,50 +44,49 @@ public class IndividualCustomerManager implements IndividualCustomerService{
 
 	@Override
 	public Result save(CreateIndividualCustomerRequest createIndividualCustomerRequest) {
-//		int max=1900;
-//		int min=0;
-//		int range = max - min + 1;
-//		int randomNumberInt = (int)(Math.random() * range) + min;
-
+		Result result = BusinessRules.run(userService.existsByEmail(createIndividualCustomerRequest.getEmail()));
+		if (result != null){
+			return result;
+		}
 		IndividualCustomer individualCustomer = modelMapperService.forRequest().map(createIndividualCustomerRequest, IndividualCustomer.class);
 		this.individualCustomerDao.save(individualCustomer);
-		return new SuccessResult("Individual customer added.");
+		return new SuccessResult(IndividualCustomerMessages.add);
 	}
 
 	@Override
 	public Result delete(DeleteIndividualCustomerRequest deleteIndividualCustomerRequest) {
+		Result result = BusinessRules.run(checkIfUserIdExists(deleteIndividualCustomerRequest.getUserId()));
+		if (result != null){
+			return result;
+		}
 		IndividualCustomer individualCustomer = modelMapperService.forRequest().map(deleteIndividualCustomerRequest, IndividualCustomer.class);
 		this.individualCustomerDao.delete(individualCustomer);
-		return new SuccessResult("Individual customer deleted.");
+		return new SuccessResult(IndividualCustomerMessages.delete);
 	}
 
 	@Override
 	public Result update(UpdateIndividualCustomerRequest updateIndividualCustomerRequest) {
+		Result result = BusinessRules.run(checkIfUserIdExists(updateIndividualCustomerRequest.getUserId()));
+		if (result != null){
+			return result;
+		}
 		IndividualCustomer individualCustomer = modelMapperService.forRequest().map(updateIndividualCustomerRequest, IndividualCustomer.class);
 		this.individualCustomerDao.save(individualCustomer);
-		return new SuccessResult("Individual customer updated.");
+		return new SuccessResult(IndividualCustomerMessages.update);
 	}
-	
-//	private Result checkUserIdDuplication(int userId) {
-//		IndividualCustomer individualCustomer = this.individualCustomerDao.getByUser_UserId(userId);
-//		if(individualCustomer != null) {
-//			return new ErrorResult("Kullanıcı numarası tekrarlayamaz.");
-//		}
-//		return new SuccessResult();
-//	}
-//	
-//	private Result checkIfUserExist(int userId) {
-//		for (User users : userDao.findAll()) {
-//			if(users.getUserId() == userId) {
-//				return new SuccessResult();
-//			}
-//		}
-//		return new ErrorResult("Kullanıcı bulunamadı.");
-//	}
 
 	@Override
 	public IndividualCustomer getCustomerByCustomerId(int customerId) {
 		IndividualCustomer individualCustomer = individualCustomerDao.getById(customerId);
 		return individualCustomer;
 	}
+
+	private Result checkIfUserIdExists(int userId){
+		IndividualCustomer individualCustomer = individualCustomerDao.getByUserId(userId);
+		if (individualCustomer == null){
+			return new ErrorResult(UserMessages.userDoesNotExist);
+		}
+		return new SuccessResult();
+	}
+
 }
